move location of log class
